installed scss and started some basic styles
