+++ client/src/Sass/base/_global.scss
@@ -1,0 +1,12 @@
+@use "../base/vars" as v;
+
+.global {
+    &__container {
+        display: "flex";
+        flex-direction: "column";
+        padding: v.$M v.$L;
+        // background-color: v.$cardBackgroundPrimary;
+        // box-shadow: v.$shadow;
+        border-radius: "2px";
+    }
+}
