+++ client/src/Sass/base/_index.scss
@@ -1,0 +1,3 @@
+@forward 'global';
+@forward 'reset';
+@forward 'vars';
