+++ client/src/Sass/components/_index.scss
@@ -1,0 +1,3 @@
+@forward "header";
+@forward "items";
+@forward "searchForm";
