+++ client/src/Sass/components/_searchForm.scss
@@ -1,0 +1,3 @@
+.searchForm {
+    border: 1px solid white;
+}
