+++ client/src/Sass/layouts/_App.scss
@@ -1,0 +1,13 @@
+@use "../base/vars" as v;
+
+.App {
+    background-color: v.$appBackground;
+    min-height: 100vh;
+    width: 100vw;
+    display: flex;
+    flex-direction: column;
+
+    &__centered {
+        @include v.flexColCenter();
+    }
+}
