+++ client/src/Sass/layouts/_index.scss
@@ -1,0 +1,1 @@
+@forward 'App';
