+++ client/src/components/item-components/ItemsContainer.js
@@ -20,19 +20,21 @@ const ItemsContainer = ({ searchedItem, items, clicked, updateList }) => {
     return (
         <div>
             <h1>Items:</h1>
-            {!clicked || targetItem === undefined ? (
-                items.map((item) => {
-                    return (
-                        <ItemCard
-                            data={item}
-                            key={item.ID}
-                            setSelectedId={setSelectedId}
-                        />
-                    );
-                })
-            ) : (
-                <ItemDetails targetItem={targetItem} />
-            )}
+            <div className="items">
+                {!clicked || targetItem === undefined ? (
+                    items.map((item) => {
+                        return (
+                            <ItemCard
+                                data={item}
+                                key={item.ID}
+                                setSelectedId={setSelectedId}
+                            />
+                        );
+                    })
+                ) : (
+                    <ItemDetails targetItem={targetItem} />
+                )}
+            </div>
         </div>
     );
 };
